+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.50.2
 - [🚀] `SelectableListX`增加`reset`扩展;
 - [🚀] 增加通用`Capturer`组件, 用于对Widget截图;
+- [🚀] `DecoratedApp`增加内建`Capturer`的选项;
 
 ## 0.50.1
 - [🚀] `decorated_app`增加`home`参数;